Agregada documentacion, y algunos métodos para vector2<float>.

+++ movimiento_recto.h
@@ -3,11 +3,17 @@
 
 namespace movimiento
 {
+/**
+ * Movimiento de linea recta.
+ */
 class mov_recto : public sub_mov
 {
 protected:
-    vel_mot actualizar_imp(util::micros_t) override;
+	vel_mot actualizar_imp(util::micros_t) override;
 public:
-	mov_recto(double);
+	/**
+	 * Cantidad a moverse para adelante.
+	 */
+	mov_recto(float);
 };
 }
